fix style and css
Various styles didn't work before the commit. The output table didn't match with the other buttons but adding the borders of various buttons i can write the right width.

the html file has been modified to include all operations in various divs. This to prevent the operation to be too distant from the numbers, giving the calculator a bad user experience (also in terms of comodity