Add events for clearing output message,operations
Creating clear and result variables and linking them to clear and result buttons, the code is now functioning as expected. The code takes a string and transforms its parts into numbers mantaining the operation and returns the result calling operate function. However the function cannot handle order of operation symbols and numbers.