Fix the multiple-digit numbers
In the execution the number that overflows the 1 digit was taken as another number. This because the split method takes the blank spaces to separate numbers and operators, and the numbers were spaced out every time the button was getting hit. This was solved deleting the space in the button content and replacing the operators with double blank spaces (left and right)