Fix the multiple floating points error
There was an error in the logic of the calculator. If i had started with an operation that wouldn't need the use of floating point numbers, then in the next operation the code would print two points on the output selector. The use of the includes() function in the floatingPoint event listener checks if the point is already in the text,thus stopping it from repeating more times